Remove id fields in the main model.  That was stupid.  Put them back in the sql wrappers where they belong.

+++ io/sql/DatabaseReader.cs
@@ -41,96 +41,96 @@ namespace XcAnalyze.Io.Sql
 
 		public Data Read ()
 		{
-			Dictionary<uint, SqlConference> conferences = ReadConferences ();
-			Dictionary<uint, Runner> runners = ReadRunners ();
-			Dictionary<uint, School> schools = ReadSchools (conferences);
-			Dictionary<uint, Affiliation> affiliations = ReadAffiliations (runners, schools);
-			Dictionary<uint, SqlMeet> meets = ReadMeets ();
-			Dictionary<uint, SqlVenue> venues = ReadVenues ();
-			Dictionary<uint, Race> races = ReadRaces (meets, venues);
-			Dictionary<uint, Performance> performances = ReadPerformances (runners, races);
+			Dictionary<int, SqlConference> conferences = ReadConferences ();
+			Dictionary<int, Runner> runners = ReadRunners ();
+			Dictionary<int, School> schools = ReadSchools (conferences);
+			Dictionary<int, Affiliation> affiliations = ReadAffiliations (runners, schools);
+			Dictionary<int, SqlMeet> meets = ReadMeets ();
+			Dictionary<int, SqlVenue> venues = ReadVenues ();
+			Dictionary<int, Race> races = ReadRaces (meets, venues);
+			Dictionary<int, Performance> performances = ReadPerformances (runners, races);
 			return new Data (new List<Affiliation>(affiliations.Values), new List<Performance>(performances.Values), new List<Race>(races.Values), new List<Runner>(runners.Values), new List<School>(schools.Values));
 		}
 		
-		Dictionary<uint, Affiliation> ReadAffiliations (Dictionary<uint, Runner> runners, Dictionary<uint, School> schools)
+		Dictionary<int, Affiliation> ReadAffiliations (Dictionary<int, Runner> runners, Dictionary<int, School> schools)
 		{
-			Dictionary<uint, Affiliation> affiliations = new Dictionary<uint, Affiliation> ();
-			uint id, runnerId, schoolId;
+			Dictionary<int, Affiliation> affiliations = new Dictionary<int, Affiliation> ();
+			int id, runnerId, schoolId;
 			Command.CommandText = "SELECT * FROM affiliations";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
-				runnerId = (uint)reader["runner_id"];
-				schoolId = (uint)reader["school_id"];
+				id = (int)(uint)reader["id"];
+				runnerId = (int)(uint)reader["runner_id"];
+				schoolId = (int)(uint)reader["school_id"];
 				affiliations.Add (id, new SqlAffiliation ((int)id, runners[runnerId], schools[schoolId], (int)reader["year"]));
 			}
 			Reader.Close ();
 			return affiliations;
 		}
 		
-		Dictionary<uint, SqlConference> ReadConferences ()
+		Dictionary<int, SqlConference> ReadConferences ()
 		{
-			Dictionary<uint, SqlConference> conferences = new Dictionary<uint, SqlConference> ();
-			uint id;
+			Dictionary<int, SqlConference> conferences = new Dictionary<int, SqlConference> ();
+			int id;
 			Command.CommandText = "SELECT * FROM conferences";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				conferences.Add (id, new SqlConference (id, (string)reader["name"], (string)reader["abbreviation"]));
 			}
 			Reader.Close ();
 			return conferences;
 		}
 		
-		Dictionary<uint, SqlMeet> ReadMeets ()
+		Dictionary<int, SqlMeet> ReadMeets ()
 		{
-			Dictionary<uint, SqlMeet> meets = new Dictionary<uint, SqlMeet> ();
-			uint id;
+			Dictionary<int, SqlMeet> meets = new Dictionary<int, SqlMeet> ();
+			int id;
 			Command.CommandText = "SELECT * FROM meets";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				meets.Add (id, new SqlMeet (id, (string)reader["name"]));
 			}
 			Reader.Close ();
 			return meets;
 		}
 		
-		Dictionary<uint, Performance> ReadPerformances (Dictionary<uint, Runner> runners, Dictionary<uint, Race> races)
+		Dictionary<int, Performance> ReadPerformances (Dictionary<int, Runner> runners, Dictionary<int, Race> races)
 		{
-			Dictionary<uint, Performance> performances = new Dictionary<uint, Performance> ();
-			uint id, runnerId, raceId;
+			Dictionary<int, Performance> performances = new Dictionary<int, Performance> ();
+			int id, runnerId, raceId;
 			Command.CommandText = "SELECT * FROM results";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
-				runnerId = (uint)reader["runner_id"];
-				raceId = (uint)reader["race_id"];
+				id = (int)(uint)reader["id"];
+				runnerId = (int)(uint)reader["runner_id"];
+				raceId = (int)(uint)reader["race_id"];
 				performances.Add (id, new SqlPerformance ((int)id, runners[runnerId], races[raceId], new Time((double)reader["time"])));
 			}
 			Reader.Close ();
 			return performances;
 		}
 		
-		Dictionary<uint, Race> ReadRaces (Dictionary<uint, SqlMeet> meets, Dictionary<uint, SqlVenue> venues)
+		Dictionary<int, Race> ReadRaces (Dictionary<int, SqlMeet> meets, Dictionary<int, SqlVenue> venues)
 		{
-			Dictionary<uint, Race> races = new Dictionary<uint, Race> ();
-			uint id;
+			Dictionary<int, Race> races = new Dictionary<int, Race> ();
+			int id;
 			SqlMeet meet;
 			SqlVenue venue;
 			Command.CommandText = "SELECT * FROM races";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				if (reader["meet_id"] is DBNull) {
 					meet = null;
 				} else {
-					meet = meets[(uint)reader["meet_id"]];
+					meet = meets[(int)(uint)reader["meet_id"]];
 				}
 				if (reader["venue_id"] is DBNull) {
 					venue = null;
 				} else {
-					venue = venues[(uint)reader["venue_id"]];
+					venue = venues[(int)(uint)reader["venue_id"]];
 				}
 				races.Add (id, SqlRace.NewInstance (id, meet, venue, new Date((DateTime)reader["date"]), Gender.FromString ((string)reader["gender"]), (int)reader["distance"]));
 			}
@@ -138,16 +138,16 @@ namespace XcAnalyze.Io.Sql
 			return races;
 		}
 		
-		Dictionary<uint, Runner> ReadRunners ()
+		Dictionary<int, Runner> ReadRunners ()
 		{
-			Dictionary<uint, Runner> runners = new Dictionary<uint, Runner> ();
-			uint id;
+			Dictionary<int, Runner> runners = new Dictionary<int, Runner> ();
+			int id;
 			string[] nicknames;
 			int? year;
 			Command.CommandText = "SELECT * FROM runners";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				if (reader["nicknames"] is DBNull) {
 					nicknames = null;
 				} else {
@@ -167,17 +167,17 @@ namespace XcAnalyze.Io.Sql
 			return runners;
 		}
 
-		Dictionary<uint, School> ReadSchools (Dictionary<uint, SqlConference> conferences)
+		Dictionary<int, School> ReadSchools (Dictionary<int, SqlConference> conferences)
 		{
-			Dictionary<uint, School> schools = new Dictionary<uint, School> ();
-			uint id;
+			Dictionary<int, School> schools = new Dictionary<int, School> ();
+			int id;
 			SqlConference conference;
 			string[] nicknames;
 			string type;
 			Command.CommandText = "SELECT * FROM schools";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ()) {
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				if (reader["nicknames"] is DBNull) {
 					nicknames = null;
 				} else {
@@ -194,7 +194,7 @@ namespace XcAnalyze.Io.Sql
 				if (reader["conference_id"] is DBNull) {
 					conference = null;
 				} else {
-					conference = conferences[(uint)reader["conference_id"]];
+					conference = conferences[(int)(uint)reader["conference_id"]];
 				}
 				schools.Add (id, SqlSchool.NewInstance (id, (string)reader["name"], nicknames, type, (bool)reader["name_first"], conference));
 			}
@@ -202,17 +202,17 @@ namespace XcAnalyze.Io.Sql
 			return schools;
 		}
 		
-		Dictionary<uint, SqlVenue> ReadVenues ()
+		Dictionary<int, SqlVenue> ReadVenues ()
 		{
-			Dictionary<uint, SqlVenue> venues = new Dictionary<uint, SqlVenue> ();
-			uint id;
+			Dictionary<int, SqlVenue> venues = new Dictionary<int, SqlVenue> ();
+			int id;
 			int? elevation;
 			string name;
 			Command.CommandText = "SELECT * FROM venues";
 			Reader = Command.ExecuteReader ();
 			while (Reader.Read ())
 			{
-				id = (uint)reader["id"];
+				id = (int)(uint)reader["id"];
 				if (reader["name"] is DBNull) 
 				{
 					name = null;
